+++ packages/enum-converter-api/src/app/schemas/schema.js
@@ -1,11 +1,11 @@
 import { ApolloServer, makeExecutableSchema } from 'apollo-server-express';
-import { convertResolvers, convertTypeDefs } from './convert';
+import { optionsResolvers, optionsTypeDefs } from './options';
 import { queryResolvers, queryTypeDef } from './query';
 import { versionResolvers, versionTypeDefs } from './version';
 
-const typeDefs = [queryTypeDef, versionTypeDefs, convertTypeDefs];
+const typeDefs = [queryTypeDef, versionTypeDefs, optionsTypeDefs];
 
-const resolvers = [queryResolvers, versionResolvers, convertResolvers];
+const resolvers = [queryResolvers, versionResolvers, optionsResolvers];
 
 const schema = makeExecutableSchema({ typeDefs, resolvers });
